Fixed stats update while reuploading same Kindle Clippings file
The issue was in the file I was using, prolly two different versions of the same book. Highlighted the same loc line but had different text. So everytime the app thought there was a new update, but the update got purged when removing redundancy yet before purging it added a quote count resulting in increase in highlight count by 1. Fixed by adding an additional condition in compareHighlights to check for timestamps of the new highlight and highlight present on cloud and only to consider the new highlight new if the time of the new highlight is greater than the time of the cloud highlight.